Normalized composer.json (#675)
* Normalized composer.json

* chore: Update composer.json and normalize formatting

The changes in composer.json include updating the "json-fix" script to also run the "composer normalize" command. This ensures that the composer.json file is properly formatted and adheres to the defined standards. Additionally, the script now applies the "composer normalize" command to all composer.json files found in the ./src directory.

Based on the recent user commits and repository commits, the commit message follows the established convention of using a prefix to indicate the type of change (`chore` for general maintenance or housekeeping).

* chore: Update composer.json script to run "composer normalize" and apply it to all composer.json files in ./src directory

* chore: Update composer.json script to run "composer normalize" and apply it to all composer.json files in ./src directory

---------

Co-authored-by: Deeka Wong <8337659+huangdijia@users.noreply.github.com>